+++ packages/venia-ui/lib/components/ARModal/ARModal.js
@@ -3,12 +3,15 @@ import { X as CloseIcon } from 'react-feather';
 
 import Icon from '../Icon';
 import { Modal } from '../Modal';
-import { Video, ARScene } from './ARViewer';
+import { Video, ARScene, Sliders } from './ARViewer';
 
 import filterModalClasses from './ARModal.css';
 
 function ARModal({ show, handleClose }) {
     const [stream, setStream] = useState(null);
+    const [position, setPosition] = useState({ x: 0, y: 10, z: 0 });
+    const [rotation, setRotation] = useState([0, 0, 0]);
+    const [element, setElement] = useState(null);
 
     const modalClass = show
         ? filterModalClasses.root_open
@@ -19,6 +22,36 @@ function ARModal({ show, handleClose }) {
         setStream(null);
         handleClose();
     }, [stream, handleClose]);
+
+    const handleXRotation = useCallback(
+        (_, newValue) => {
+            const xRotation = Number(newValue);
+            if (xRotation) {
+                setRotation([xRotation, rotation[1], rotation[2]]);
+            }
+        },
+        [rotation, setRotation]
+    );
+
+    const handleYRotation = useCallback(
+        (_, newValue) => {
+            const yRotation = Number(newValue);
+            if (yRotation) {
+                setRotation([rotation[0], yRotation, rotation[2]]);
+            }
+        },
+        [rotation, setRotation]
+    );
+
+    const handleZRotation = useCallback(
+        (_, newValue) => {
+            const zRotation = Number(newValue);
+            if (zRotation) {
+                setRotation([rotation[0], rotation[1], zRotation]);
+            }
+        },
+        [rotation, setRotation]
+    );
 
     return show ? (
         <Modal>
@@ -33,8 +66,21 @@ function ARModal({ show, handleClose }) {
                         </button>
                     </div>
                     <Video registerStream={setStream}>
-                        <ARScene />
+                        <ARScene
+                            setElement={setElement}
+                            setPosition={setPosition}
+                            setRotation={setRotation}
+                            element={element}
+                            rotation={rotation}
+                            position={position}
+                        />
                     </Video>
+                    <Sliders
+                        handleXRotation={handleXRotation}
+                        handleYRotation={handleYRotation}
+                        handleZRotation={handleZRotation}
+                        element={element}
+                    />
                 </div>
             </aside>
         </Modal>
